Organize project files
